+++ src/templates/table.ts
@@ -10,7 +10,12 @@ export class {{ name }} extends Entity {
     }
 
     {{{table-of-fields 'buf' fields}}}
-    return new {{ name }}(fields)
+
+    const entity = new {{ name }}(fields)
+    if (!entity.verify(buf)) {
+      throw new Error('Invalid binary data')
+    }
+    return entity
   }
 
   {{#each fields}}
@@ -29,7 +34,7 @@ export class {{ name }} extends Entity {
   }
 
   get size (): number {
-    let headerLength = uint32Length + uint32Length * {{ fields.length }}
+    const headerLength = uint32Length + uint32Length * {{ fields.length }}
     let bodyLength = 0
     {{#each fields}}
       {{#if (is-equal type 'Buffer')}}
@@ -43,7 +48,7 @@ export class {{ name }} extends Entity {
 
   get offsets (): number[] {
     let start = uint32Length * (1 + {{fields.length}})
-    let offsets: number[] = []
+    const offsets: number[] = []
 
     {{#each fields}}
     offsets.push(start)
@@ -64,10 +69,11 @@ export class {{ name }} extends Entity {
   set {{ name }} (val: {{type}}) {
     this._{{ name }} = val
   }
+
   {{/each}}
 
   toBuffer (): Buffer {
-    let header = Buffer.alloc(uint32Length * (1 + {{fields.length}}))
+    const header = Buffer.alloc(uint32Length * (1 + {{fields.length}}))
     header.writeUInt32LE(this.size)
     let start = uint32Length
     this.offsets.forEach(item => {
@@ -75,7 +81,7 @@ export class {{ name }} extends Entity {
       start += uint32Length
     })
 
-    let bufs = []
+    const bufs = []
     {{#each fields}}
       {{#if (is-equal type 'Buffer')}}
         bufs.push(this._{{ name }})
@@ -92,7 +98,7 @@ export class {{ name }} extends Entity {
   }
 
   toRawData (): Buffer {
-    let bufs = [
+    const bufs = [
       {{#each fields}}
         {{#if (is-equal type 'Buffer')}}
           this._{{ name }},